+++ Main.py
@@ -17,7 +17,6 @@ from OverworldGlitchRules import create_owg_connections
 from PotShuffle import shuffle_pots, shuffle_pot_switches
 from Regions import create_regions, create_shops, mark_light_dark_world_regions, create_dungeon_regions, adjust_locations
 from OverworldShuffle import link_overworld, create_dynamic_exits
-from EntranceShuffle import link_entrances
 from Rom import patch_rom, patch_race_rom, apply_rom_settings, LocalRom, JsonRom, get_hash_string
 from Doors import create_doors
 from DoorShuffle import link_doors, connect_portal, link_doors_prep
@@ -39,7 +38,7 @@ from source.enemizer.DamageTables import DamageTable
 from source.enemizer.Enemizer import randomize_enemies
 from source.rom.DataTables import init_data_tables
 
-version_number = '1.4.1.11'
+version_number = '1.4.1.12'
 version_branch = '-u'
 __version__ = f'{version_number}{version_branch}'
 
@@ -249,10 +248,7 @@ def main(args, seed=None, fish=None):
     for player in range(1, world.players + 1):
         link_overworld(world, player)
         create_dynamic_exits(world, player)
-        if world.experimental[player] or world.shuffle[player] in ['lite', 'lean', 'swapped'] or world.shuffletavern[player] or (world.customizer and world.customizer.get_entrances()):
-            link_entrances_new(world, player)
-        else:
-            link_entrances(world, player)
+        link_entrances_new(world, player)
         if world.logic[player] in ('nologic', 'hybridglitches'):
             create_hybridmajor_connectors(world, player)