+++ test/inverted/TestInverted.py
@@ -3,12 +3,13 @@ from DoorShuffle import link_doors
 from Doors import create_doors
 from Dungeons import create_dungeons, get_dungeon_item_pool
 from OverworldShuffle import link_overworld
-from EntranceShuffle import link_entrances
 from ItemList import generate_itempool, difficulties
 from Items import ItemFactory
 from Regions import create_regions, mark_light_dark_world_regions, create_dungeon_regions, create_shops
 from RoomData import create_rooms
 from Rules import set_rules
+
+from source.overworld.EntranceShuffle2 import link_entrances_new
 from test.TestBase import TestBase
 
 
@@ -25,7 +26,7 @@ class TestInverted(TestBase):
         create_rooms(self.world, 1)
         create_dungeons(self.world, 1)
         link_overworld(self.world, 1)
-        link_entrances(self.world, 1)
+        link_entrances_new(self.world, 1)
         link_doors(self.world, 1)
         generate_itempool(self.world, 1)
         self.world.required_medallions[1] = ['Ether', 'Quake']