+++ test/vanilla/TestVanilla.py
@@ -3,7 +3,7 @@ from DoorShuffle import link_doors
 from Doors import create_doors
 from Dungeons import create_dungeons, get_dungeon_item_pool
 from OverworldShuffle import link_overworld
-from EntranceShuffle import link_entrances
+from source.overworld.EntranceShuffle2 import link_entrances_new
 from ItemList import difficulties, generate_itempool
 from Items import ItemFactory
 from Regions import create_regions, create_dungeon_regions, create_shops, mark_light_dark_world_regions
@@ -25,7 +25,7 @@ class TestVanilla(TestBase):
         create_rooms(self.world, 1)
         create_dungeons(self.world, 1)
         link_overworld(self.world, 1)
-        link_entrances(self.world, 1)
+        link_entrances_new(self.world, 1)
         link_doors(self.world, 1)
         generate_itempool(self.world, 1)
         self.world.required_medallions[1] = ['Ether', 'Quake']